+++ app/containers/ContactsControl/reducer.js
@@ -27,14 +27,19 @@ function contactsControlReducer(state = initialState, action) {
     }
     case SET_SEARCH_RESULTS: {
       let results = state.get('results');
+      let validResultsCount = action.response.count;
       if (action.response.page === state.get('nextPage')) {
-        action.response.results.forEach((result) => {
-          results = results.push(fromJS(result));
+        action.response.results.forEach((newResult) => {
+          if (newResult) { // search returns null for contact immediately after it's been deleted
+            results = results.push(fromJS(newResult));
+          } else {
+            validResultsCount -= 1;
+          }
         });
         return state
           .set('results', results)
           .set('nextPage', action.response.page + 1)
-          .set('resultsCount', action.response.count);
+          .set('resultsCount', validResultsCount);
       } else {
         return state;
       }